# internal/docs/swagger.yaml
basePath: /
definitions:
  api.AnalyzeCounterDetailRequest:
    properties:
      counterHeroId:
        example: diggie
        type: string
      language:
        enum:
        - en
        - id
        example: en
        type: string
      targetHeroId:
        example: tigreal
        type: string
    type: object
  api.AnalyzeCounterDetailResponse:
    properties:
      conditions:
        items:
          type: string
        type: array
      confidence:
        example: 85
        type: integer
      counterHeroId:
        example: diggie
        type: string
      evidenceIds:
        items:
          type: string
        type: array
      failureCases:
        items:
          type: string
        type: array
      score:
        example: 90
        type: integer
      source:
        example: ai
        type: string
      strengths:
        items:
          type: string
        type: array
      summary:
        type: string
      targetHeroId:
        example: tigreal
        type: string
    type: object
  api.AnalyzeCounterScoreRequest:
    properties:
      language:
        enum:
        - en
        - id
        example: en
        type: string
      targetHeroId:
        example: tigreal
        type: string
    type: object
  api.AnalyzeCounterScoreResponse:
    properties:
      recommendations:
        items:
          $ref: '#/definitions/api.CounterScoreRecommendation'
        type: array
      source:
        example: ai
        type: string
      targetHeroId:
        example: tigreal
        type: string
    type: object
  api.AnalyzeSynergyDetailRequest:
    properties:
      anchorHeroId:
        example: tigreal
        type: string
      language:
        enum:
        - en
        - id
        example: en
        type: string
      synergyHeroId:
        example: pharsa
        type: string
    type: object
  api.AnalyzeSynergyDetailResponse:
    properties:
      anchorHeroId:
        example: tigreal
        type: string
      conditions:
        items:
          type: string
        type: array
      confidence:
        example: 85
        type: integer
      evidenceIds:
        items:
          type: string
        type: array
      failureCases:
        items:
          type: string
        type: array
      score:
        example: 90
        type: integer
      source:
        example: ai
        type: string
      strengths:
        items:
          type: string
        type: array
      summary:
        type: string
      synergyHeroId:
        example: pharsa
        type: string
    type: object
  api.AnalyzeSynergyScoreRequest:
    properties:
      anchorHeroId:
        example: tigreal
        type: string
      language:
        enum:
        - en
        - id
        example: en
        type: string
    type: object
  api.AnalyzeSynergyScoreResponse:
    properties:
      anchorHeroId:
        example: tigreal
        type: string
      recommendations:
        items:
          $ref: '#/definitions/api.SynergyScoreRecommendation'
        type: array
      source:
        example: ai
        type: string
    type: object
...
  api.CounterScoreRecommendation:
    properties:
      confidence:
        example: 85
        type: integer
      counterHeroId:
        example: diggie
        type: string
      rank:
        example: 1
        type: integer
      score:
        example: 90
        type: integer
    type: object
  api.ErrorBody:
    properties:
      code:
        example: hero_not_found
        type: string
      message:
        example: Hero was not found in the dataset.
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorBody'
    type: object
...
  api.SynergyScoreRecommendation:
    properties:
      confidence:
        example: 85
        type: integer
      rank:
        example: 1
        type: integer
      score:
        example: 90
        type: integer
      synergyHeroId:
        example: pharsa
        type: string
    type: object
...
info:
  contact: {}
  description: Serves the MLBB hero, counter, and synergy dataset from Postgres and
    returns AI-scored counter recommendations for frontend consumption.
  title: MLBB Analyzer Service
  version: "1.0"
paths:
...
  /api/counters/analyze-detail:
    post:
      consumes:
      - application/json
      description: Returns summary, strengths, conditions, and failure cases for one
        target/counter pair, grounded only in authored proof evidence.
      parameters:
      - description: Target hero, counter hero, and output language
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AnalyzeCounterDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AnalyzeCounterDetailResponse'
        "404":
          description: target_hero_not_found, counter_hero_not_found, counter_matchup_not_found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: invalid_request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "429":
          description: rate_limit_exceeded
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: ai_provider_error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: rate_limit_unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "504":
          description: ai_provider_not_configured
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Explain one counter matchup in detail
      tags:
      - counters
  /api/counters/analyze-score:
    post:
      consumes:
      - application/json
      description: Returns AI-produced score and confidence for every counter matchup
        of the target hero, ranked best-first. Cached results do not consume rate-limit
        quota.
      parameters:
      - description: Target hero and output language
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AnalyzeCounterScoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AnalyzeCounterScoreResponse'
        "404":
          description: target_hero_not_found, counter_data_not_found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: invalid_request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "429":
          description: rate_limit_exceeded
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: ai_provider_error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: rate_limit_unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "504":
          description: ai_provider_not_configured
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Score all counter matchups for one target hero
      tags:
      - counters
...
  /api/synergies/analyze-detail:
    post:
      consumes:
      - application/json
      description: Returns summary, strengths, conditions, and failure cases for one
        anchor/synergy pair, grounded only in authored proof evidence.
      parameters:
      - description: Anchor hero, synergy hero, and output language
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AnalyzeSynergyDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AnalyzeSynergyDetailResponse'
        "404":
          description: anchor_hero_not_found, synergy_hero_not_found, synergy_matchup_not_found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: invalid_request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "429":
          description: rate_limit_exceeded
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: ai_provider_error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: rate_limit_unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "504":
          description: ai_provider_not_configured
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Explain one synergy pairing in detail
      tags:
      - synergies
  /api/synergies/analyze-score:
    post:
      consumes:
      - application/json
      description: Returns AI-produced score and confidence for every synergy pairing
        of the anchor hero, ranked best-first. Cached results do not consume rate-limit
        quota.
      parameters:
      - description: Anchor hero and output language
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AnalyzeSynergyScoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AnalyzeSynergyScoreResponse'
        "404":
          description: anchor_hero_not_found, synergy_data_not_found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: invalid_request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "429":
          description: rate_limit_exceeded
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: ai_provider_error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: rate_limit_unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "504":
          description: ai_provider_not_configured
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Score all synergy pairings for one anchor hero
      tags:
      - synergies
...
swagger: "2.0"
...